[Rename] 컴포넌트명 수정
