+++ components/exchange/chart/order/History.jsx
@@ -29,7 +29,7 @@ const CancelButton = styled(Button)(() => ({
   },
 }));
 
-export default function ModalHistory({ value, orders, removeOrder }) {
+export default function History({ value, orders, removeOrder }) {
   const [selectedValue, setSelectedValue] = useState('a');
   const [open, setOpen] = useState(false); // 좌측 하단 알림 오픈
